+++ src/constants.js
@@ -12,6 +12,7 @@ const constants = {
  * @type {ESSessionConstants} Defines all the module session storage static constants.
  */
 const sessionConstants = {
+   appState: `${constants.moduleId}.app-state`,
    themeDarkMode: `${constants.moduleId}.theme.dark`
 };
 
@@ -19,6 +20,8 @@ export { constants, sessionConstants };
 
 /**
  * @typedef {object} ESSessionConstants
+ *
+ * @property {string} appState - Stores the current app state in /src/view/app-state demo
  *
  * @property {string} themeDarkMode - Indicates whether dark mode is enabled.
  */