+++ src/view/editor/content-editable/ContentEditableShell.svelte
@@ -14,16 +14,17 @@
     *
     * The simplest example is:
     *
-    * <TJSProseMirror options={{document: <doc>, fieldName: 'some.data.path'}} />
+    * <TJSContentEdit options={{document: <doc>, fieldName: 'some.data.path'}} />
     *
-    * The following options data is commented out as it is set to a specific document for the DnD5e system.
+    * The following options data is mostly commented out. For the most part though the values shown for configuration
+    * show the default values _or_ in some cases the type of data that you need to set.
     */
    const options = {
-      // document: game.items.get('cUZEGVZdhr6G9QcM'),   // An item to edit description; note: replace w/ valid doc.
+      // document: game.items.get('OYHnejmJO2fSlQDi'),   // An item to edit description; note: replace w/ valid doc.
       // fieldName: 'system.description.value',          // Path to data in `a.b.c`; note: this is a v10 field name.
 
       // button: true      // Show edit button to launch editor when hovered; when false editor is open by default.
-      // classes: ['foo', 'bar'],   // Adds additional classes to `.tjs-contenteditable` element.
+      // classes: ['foo', 'bar'],   // Adds additional classes to `.tjs-editor` element.
       // clickToEdit: false,  // Clicking editor content initializes the editor; hides the edit button.
       // editable: true,   // Enable / disable editing
       // DOMPurify,        // You can pass DOMPurify from `@typhonjs-fvtt/runtime/dompurify though ProseMirror does
@@ -33,6 +34,7 @@
       // enrichContent: true  // The default is true, but if you set it to false content is not enriched.
 
       // initialSelection: 'start', // The initial selection / cursor position: 'all', 'end', or 'start'.
+      // maxCharacterLength: 25, // Limits input to the number of characters; pasting is text only / any HTML stripped.
 
       // preventEnterKey: false, // Prevents <enter> key / new lines.
       // preventPaste: false,    // Prevents pasting.
@@ -62,8 +64,9 @@
                    bind:content
                    bind:enrichedContent
                    on:editor:cancel={() => console.log('! event - editor:cancel')}
+                   on:editor:document:deleted={() => console.log('! event - editor:document:deleted')}
                    on:editor:enrichedContent={(event) => console.log(`! event - editor:enrichedContent - ${event.detail.enrichedContent}`)}
                    on:editor:save={(event) => console.log(`! event - editor:save - ${event.detail.content}`)}
                    on:editor:start={() => console.log('! event - editor:start')} />
-                   <!-- You can subscribe to the above events if desired -->
+                   <!-- Optionally, you can subscribe to the above events if desired -->
 </ApplicationShell>