+++ src/view/editor/prosemirror/ProseMirrorShell.svelte
@@ -3,7 +3,7 @@
 
    import { TJSProseMirror }     from '@typhonjs-fvtt/svelte-standard/component';
 
-   // If you are developing ProseMirror plugins import from this path to enable PM debugging tooling.
+   // If you are developing ProseMirror plugins import from this path to enable PM inspector / debug tooling.
    // import { TJSProseMirror }     from '@typhonjs-fvtt/svelte-standard/component/dev';
 
    // Not always necessary, but you can use DOMPurify to sanitize user input client side.
@@ -19,15 +19,16 @@
     *
     * <TJSProseMirror options={{document: <doc>, fieldName: 'some.data.path'}} />
     *
-    * The following options data is commented out as it is set to a specific document for the DnD5e system.
+    * The following options data is mostly commented out. For the most part though the values shown for configuration
+    * show the default values _or_ in some cases the type of data that you need to set.
     */
    const options = {
-      // document: game.items.get('cUZEGVZdhr6G9QcM'),   // An item to edit description; note: replace w/ valid doc.
+      // document: game.items.get('OYHnejmJO2fSlQDi'),   // An item to edit description; note: replace w/ valid doc.
       // fieldName: 'system.description.value',          // Path to data in `a.b.c`; note: this is a v10 field name.
       // collaborate: false,                             // Enables collaboration; requires document.
 
       // button: true      // Show edit button to launch editor when hovered; when false editor is open by default.
-      // classes: ['foo', 'bar'],   // Adds additional classes to `.editor` element.
+      // classes: ['foo', 'bar'],   // Adds additional classes to `.tjs-editor` element.
       // clickToEdit: false,  // Clicking editor content initializes the editor; hides the edit button.
       // DOMPurify,        // You can pass DOMPurify from `@typhonjs-fvtt/runtime/dompurify though ProseMirror does
                            // essential client side sanitation; IE stripping `<script>` tags, etc.
@@ -57,8 +58,9 @@
                    bind:content
                    bind:enrichedContent
                    on:editor:cancel={() => console.log('! event - editor:cancel')}
+                   on:editor:document:deleted={() => console.log('! event - editor:document:deleted')}
                    on:editor:enrichedContent={(event) => console.log(`! event - editor:enrichedContent - ${event.detail.enrichedContent}`)}
                    on:editor:save={(event) => console.log(`! event - editor:save - ${event.detail.content}`)}
                    on:editor:start={() => console.log('! event - editor:start')} />
-                   <!-- You can subscribe to the above events if desired -->
+                   <!-- Optionally, you can subscribe to the above events if desired -->
 </ApplicationShell>