+++ src/view/header-buttons/HeaderButtonsAppShell.svelte
@@ -1,11 +1,10 @@
 <script>
-   import { alwaysBlur }         from '@typhonjs-fvtt/runtime/svelte/action';
    import { getContext }         from 'svelte';
    import { ApplicationShell }   from '@typhonjs-fvtt/runtime/svelte/component/core';
 
    export let elementRoot = void 0;
 
-   const application = getContext('external').application;
+   const application = getContext('#external').application;
 
    // Provides two-way bindings to application options. By using the stores external updates will be received here.
    const headerButtonNoClose = application.reactive.storeAppOptions.headerButtonNoClose;
@@ -21,9 +20,9 @@
       <br>
       <div class=container>
          Options:
-         <label><input use:alwaysBlur type=checkbox bind:checked={$headerButtonNoClose}> No Close Button</label>
-         <label><input use:alwaysBlur type=checkbox bind:checked={$headerButtonNoLabel}> No Button Labels</label>
-         <label><input use:alwaysBlur type=checkbox bind:checked={$headerNoTitleMinimized}> No Title (Minimized)</label>
+         <label><input type=checkbox bind:checked={$headerButtonNoClose}> No Close Button</label>
+         <label><input type=checkbox bind:checked={$headerButtonNoLabel}> No Button Labels</label>
+         <label><input type=checkbox bind:checked={$headerNoTitleMinimized}> No Title (Minimized)</label>
       </div>
    </main>
 </ApplicationShell>