+++ src/view/position/box/PositionBoxApplication.js
@@ -5,11 +5,6 @@ import BoxLayer               from './BoxLayer.svelte';
 export default class PositionBoxApplication extends SvelteApplication
 {
    /**
-    * @inheritDoc
-    */
-   constructor(options = {}) { super(options); }
-
-   /**
     * Default Application options
     *
     * @returns {object} options - Application options.
@@ -19,6 +14,7 @@ export default class PositionBoxApplication extends SvelteApplication
    {
       return foundry.utils.mergeObject(super.defaultOptions, {
          id: 'position-box',
+         classes: ['tjs-essential-svelte-esm'],
          title: 'Position (Box)',
          width: 860,
          height: 800,