+++ src/view/position/carousel/PositionCarouselApp.js
@@ -5,11 +5,6 @@ import CarouselAppShell       from './CarouselAppShell.svelte';
 export default class PositionBoxApplication extends SvelteApplication
 {
    /**
-    * @inheritDoc
-    */
-   constructor(options = {}) { super(options); }
-
-   /**
     * Default Application options
     *
     * @returns {object} options - Application options.
@@ -19,6 +14,7 @@ export default class PositionBoxApplication extends SvelteApplication
    {
       return foundry.utils.mergeObject(super.defaultOptions, {
          id: 'position-carousel',
+         classes: ['tjs-essential-svelte-esm'],
          title: 'Position (Carousel)',
          width: 775,
          height: 420,