+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
+## Release 0.0.15 (major update)
+- Updated @typhonjs-fvtt/runtime to ['0.0.22'](https://github.com/typhonjs-fvtt-lib/typhonjs/blob/main/CHANGELOG.md#release-0022).
+- Updated @typhonjs-fvtt/svelte-standard to ['0.0.18'](https://github.com/typhonjs-fvtt-lib/svelte-standard/blob/main/CHANGELOG.md#release-0018).
+- Finished editor component demos.
+
 ## Release 0.0.14 (major update)
 - Updated @typhonjs-fvtt/runtime to ['0.0.20'](https://github.com/typhonjs-fvtt-lib/typhonjs/blob/main/CHANGELOG.md#release-0020).
 - Updated @typhonjs-fvtt/svelte-standard to ['0.0.16'](https://github.com/typhonjs-fvtt-lib/svelte-standard/blob/main/CHANGELOG.md#release-0016).